FLEX-34306 - [BlazeDS] Make the BlazeDS build run on Windows machines - Added some mkdir commands to the ANT Build.java - Did some fine-tuning to resolve some compile errors
FLEX-34307 - [BlazeDS] Adjust the Maven build to be able to deploy SNAPSHOTs and releases to Apaches Maven repo.
- Adjusted the poms to be able to build and theoretically deploy to Apaches Maven repo.

FLEX-34308 - [BlazeDS] Make all tests run in a Maven build without any need to manually start and deploy BlazeDS to an external Server.
- Moved all tests to a new testsuite module
- Implemented a Utility class (TestServer), which starts an embedded Jetty and deploys the MessageBrokerServlet which is configured to run the needed services.

FLEX-34309 - [BlazeDS] Implement a way to configure a BlazeDS using a configuration stored in the applications Classpath
- Implemented a ClasspathResourceResolver to read a BlazeDS configuration from classpath instead of the filesystem.
- Updated the FlexConfigurationManager to use this ResourceResolver when encountering a config path starting with "classpath:"